pyproject.toml
[project]
name = "movie-warp"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = "~=3.13"
dependencies = ["bedrockpy (>=1.0.1,<2.0.0)", "moviepy (>=2.2.1,<3.0.0)", "pillow (>=11.2.1,<12.0.0)", "rgbmatch (>=0.0.1,<0.0.2)", "powercli-python (==0.1.1)"]
classifiers = ["Private :: Do Not Upload"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.16.1"

